+++ lib/aqua_ad_widget.dart
@@ -7,3 +7,4 @@ library aqua_ad_widget;
 
 export 'src/widgets/aqua_ad_widget.dart';
 export 'src/config/aqua_config.dart';
+export 'src/config/aqua_settings.dart';